refactor: extract trace writer sampling helpers
